==> 02_knit_me_summarize_data.Rmd <==
---
title: "Summary of data processing"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(DiagrammeR)
library(tidyverse)
library(htmltools)
dfs <- readRDS("dataframes.RDS")
```


### Input Dataframes:

Here is a brief description of the input dataframes (column names, data type, number of unique values and 10 example values).


```{r, echo=FALSE, results='asis'}
print_kable <- function(names, tbbl){
  cat(HTML(paste0("<h6> Input ",names,"</h6>")))
  tbbl%>%
    knitr::kable(escape = TRUE)%>%
    kableExtra::kable_styling("striped", position = "left", font_size = 7, full_width=FALSE)%>%
    print()
}

input_list <- dfs%>%
  filter(df_names %in% c("wages_cleaned",
                        "jo_raw",
                        "ind_char_raw",
                        "employment_raw",
                        "ds_raw",
                        "noc_mappings_raw",
                        "education_occupation_raw",
                        "hoo"))%>%
  deframe()

walk2(names(input_list), input_list, print_kable)
```

### Description of code:

##### Clean_JO.csv

```{r, echo=FALSE}
mermaid("
graph LR

1(JO single variables.csv)
2(Industry characteristics.xlsx)
3(Emp single variables.csv)

1 --> jo_raw
2 --> ind_char_raw
3 --> employment_raw

jo_raw --> long
employment_raw --> long
ind_char_raw --> jo_employment

long --> jo_employment
jo_employment --> Clean_JO.csv

", height = '100%', width = '100%')
```

Dataframes `jo_raw` and `employment_raw` contain LMO64 (plus total) level data for all NOC categories (plus total).  Bind these dataframes together as `long`, discarding the aggregate values.  Dataframe `ind_char_raw` contains columns `industry_code` and `aggregate_industry`.  We join `ind_char_raw` with `long` by `industry` to get dataframe `jo_employment`, retaining columns `date`, `noc`, `description`, `employment`, `industry`, `aggregate_industry`, `geographic_area`, `job_openings`, `expansion_demand`, `replacement_demand`, `deaths`, `retirements` and `industry_code`.

##### Supply_cleaned.csv

```{r, echo=FALSE}
mermaid("
graph LR

1(JO single variables.csv)
2(DS single variables.csv)

1 --> jo_raw
2 --> ds_raw

jo_raw --> jo_all_industries
jo_all_industries --> ds_and_jo
ds_raw --> ds_and_jo
ds_and_jo --> Supply_cleaned.csv
", height = '100%', width = '100%')
```

Filter dataframe `jo_raw` retaining only the "all_industry" aggregate level data. Dataframe `ds_raw` also contains "all_industry" aggregate level data on demand and supply: Row bind these dataframes together to form dataframe `ds_and_jo`, pivot wider and keep columns `date`, 
         `geographic_area`, 
         `noc`, 
         `description`, 
         `industry`, 
         `job_openings`, 
         `young_people_starting_work`, 
         `immigrants`, 
         `migrants_from_other_provinces`,
         `additional_supply_requirement`,
         `labour_force_exits` and
         `net_change_in_labour_force`.

##### Occupations_regional.csv

```{r, echo=FALSE}
mermaid("
graph LR

1(JO single variables.csv)
2(Industry characteristics.xlsx)
3(Emp single variables.csv)
4(NOC Mappings.csv)
5(Occupation characteristics.xlsx)

1 --> jo_raw
2 --> ind_char_raw
3 --> employment_raw
4 --> noc_mappings_raw
5 --> education_occupation_raw

noc_mappings_raw --> noc_major_group
noc_mappings_raw --> noc_minor_group
noc_mappings_raw --> noc_unit
education_occupation_raw --> education_occupation
jo_employment --> jobs_employment
education_occupation --> jobs_employment
jobs_employment --> occupation
occupation --> noccupation
noc_broad_occ --> noccupation
noc_major_group--> noccupation
noc_minor_group--> noccupation
noc_unit--> noccupation
noccupation --> Occupations_regional.csv
long --> jo_employment
ind_char_raw --> jo_employment
jo_raw --> long
employment_raw --> long
noc_mappings_raw --> noc_broad_occ
", height = '100%', width = '100%')
```

Dataframe `education_occupation` has columns `noc`, `noc1`, `noc2`, `noc3`, `education_typical_background` for all noc other than `#t`. Join `jo_employment` and `education_occupation` to create `jobs_employment`.  Use `jobs_employment` as an input to calculate `employment`, `expansion`, `replacement` and `job_openings` resulting in dataframe `occupation`. From dataframe `noc_mappings_raw` we create mapping dataframes from the nocs to their description, and then join with `occupation`, creating dataframe `noccupation` with columns `noc`, `description`, `geographic_area`, `noc1`, `noc2`, `noc3`, `education_typical_background`, `Employment year1`, `Expansion year1-year3`, `Replacement year1-year3`, `Job Openings year1-year3`, `noc1_description`, `noc2_description`, `noc3_description`, `noc4_description`.


##### Jobs_and_industry.csv 
```{r, echo=FALSE}
mermaid("
graph LR
1(Industry characteristics.xlsx)
2(JO single variables.csv)
3(Emp single variables.csv)
4(NOC Mappings.csv)
5(Occupation characteristics.xlsx)

1 --> ind_char_raw
2 --> jo_raw
3 --> employment_raw
4 --> noc_mappings_raw
5 --> education_occupation_raw

noc_unit --> noccupation
ind_char_raw --> ind_char2
long --> jo_employment
noccupation --> noc_mappings2
ind_char2 --> jobs_and_industry
jo_employment --> jobs_and_industry
jobs_and_industry --> jobs_industry_noc
noc_mappings2 --> jobs_industry_noc
jobs_industry_noc --> Jobs_and_Industry.csv
employment_raw --> long
ind_char_raw --> jo_employment
education_occupation_raw --> education_occupation
noc_mappings_raw --> noc_broad_occ
noc_mappings_raw --> noc_major_group
noc_mappings_raw --> noc_minor_group
noc_mappings_raw --> noc_unit
education_occupation --> jobs_employment
jo_employment --> jobs_employment
jobs_employment --> occupation
jo_raw --> long
occupation --> noccupation
noc_broad_occ --> noccupation
noc_major_group--> noccupation
noc_minor_group--> noccupation
", height = '100%', width = '100%')
```

Dataframe `ind_char2` has columns `industry_code`, `industry`, `naics_definition`, `agg_industry_code`, `aggregate_industry`, `sector_code`, `sector`.  Join `ind_char2` with `jo_employment` to create `jobs_and_industry`.  From `noccupation` get all observed nocs and call it `noc_mappings2`.  Join `jobs_and_industry` with `noc_mappings2` to create dataframe `jobs_industry_noc` with columns `date`, `noc`, `description`, `employment`, `industry`, `aggregate_industry`, `geographic_area`, `job_openings`, `expansion_demand`, `replacement_demand`, `deaths`, `retirements`, `industry_code`, `naics_definition`, `sector_code`, `sector`, `noc1`, `noc2`, `noc3`, `noc1_description`, `noc2_description`, `noc3_description`, `noc4_description`. 


##### Employment_Growth_Rates.csv 
```{r, echo=FALSE}
mermaid("
graph LR

1(JO single variables.csv)
2(Industry characteristics.xlsx)
3(Emp single variables.csv)
4(Occupation characteristics.xlsx)
5(NOC Mappings.csv)

1 --> jo_raw
2 --> ind_char_raw
3 --> employment_raw
4 --> education_occupation_raw
5 --> noc_mappings_raw

noc_mappings_raw --> noc_broad_occ
noc_mappings_raw --> noc_major_group
noc_mappings_raw --> noc_minor_group
noc_mappings_raw --> noc_unit
jobs_employment --> occupation
occupation --> noccupation
noc_broad_occ --> noccupation
noc_major_group--> noccupation
noc_minor_group--> noccupation
noc_unit--> noccupation
jo_raw --> long
employment_raw --> long
ind_char_raw --> jo_employment
long --> jo_employment
ind_char_raw --> ind_char2
noccupation --> noc_mappings2
education_occupation_raw --> education_occupation
jo_employment --> jobs_and_industry
ind_char2 --> jobs_and_industry
jobs_and_industry --> jobs_industry_noc
noc_mappings2 --> jobs_industry_noc
education_occupation --> jobs_employment
jo_employment --> jobs_employment
jobs_industry_noc --> mapping
jobs_employment --> by_aggregated_industry
jobs_employment --> by_individual_industry
mapping --> by_individual_industry
by_individual_industry --> individual_industry_agg_industry
by_aggregated_industry --> individual_industry_agg_industry
individual_industry_agg_industry --> Employment_Growth_Rates.csv
", height = '100%', width = '100%')
```

Get all unique combinations of `industry` and `aggregate_industry` from `jobs_industry_noc` and call the dataframe `mapping`.  Take dataframe `jobs_employment` and aggregate by `industry`, calculate some measures then join with `mapping` to create dataframe `by_individual_industry`. Take dataframe `jobs_employment` and aggregate by `aggregate_industry` and calculate some measures to create dataframe `by_aggregate_industry`.  Bind the rows of  `by_individual_industry` and `by_individual_industry` to create dataframe `individual_industry_agg_industry` with columns `level_value`, `geographic_area`, `variable`, `value`, `level`, `aggregate_industry`.


##### occ_characteristics_wage.csv

```{r, echo=FALSE}
mermaid("
graph LR

1(JO single variables.csv) 
2(Emp single variables.csv) 
3(Industry characteristics.xlsx) 
4(2020Preliminary wages.xlsx)
5(Occupation characteristics.xlsx)

1 --> jo_raw
2 --> employment_raw
3 --> ind_char_raw
4 --> wages_raw
5 --> education_occupation_raw

jo_raw --> long
employment_raw -->long
ind_char_raw --> jo_employment
long --> jo_employment
wages_raw --> wages_cleaned
education_occupation_raw --> education_occupation
education_occupation_raw --> occ_group
jo_employment --> jobs_employment
education_occupation --> jobs_employment
wages_cleaned --> group_and_wages
occ_group --> group_and_wages
education_occupation_raw --> occ_characteristics
jobs_employment --> j_openings
jobs_employment --> emp
group_and_wages --> group_wages_characteristics
occ_characteristics --> group_wages_characteristics
j_openings --> group_wages_characteristics
emp --> group_wages_characteristics
group_wages_characteristics --> occ_characteristics_wage.csv
", height = '100%', width = '100%')
```

Dataframe `wages_cleaned` contains wage distribution data for combinations of `noc` and `geographic_area`.  Dataframe `occ_group` contains occupation group categories.  Join `wages_cleaned` and `occ_group` to form dataframe `group_and_wages`. Dataframe `occ_characteristics` contains columns  
`noc`, `occupation_title`, `typical_education`, `alternative_education`, `interest1`, `interest2`, `interest3`, `skill1`, `skill2`, `skill3`, `interests`, `top_3_skills_and_competencies`.  From dataframe `jobs_employment` get the sum of job openings and call it `j_openings`.   From dataframe `jobs_employment` get the sum of employment and call it `emp`. Join `group_and_wages` and `occ_characteristics` by `noc`, and then join the resulting dataframe to `j_openings` and `emp` by both `noc` and `region`.  




### Output dataframe summaries:

Here is a brief description of the output dataframes (column names, data type, number of unique values and 10 example values.)


```{r, echo=FALSE, results='asis'}
print_kable <- function(names, tbbl){
  cat(HTML(paste0("<h6> Output ",names,"</h6>")))
  tbbl%>%
    knitr::kable(escape = TRUE)%>%
    kableExtra::kable_styling("striped", position = "left", font_size = 7, full_width=FALSE)%>%
    print()
}

out_list <- dfs%>%
  filter(df_names %in% c("jo_employment",
                        "ds_and_jo",
                        "noccupation",
                        "jobs_industry_noc",
                        "individual_industry_agg_industry",
                        "group_wages_characteristics"))%>%
  deframe()

walk2(names(out_list), out_list, print_kable)
```
